% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4.R
\docType{methods}
\name{biofluids,HmdbEntry-method}
\alias{biofluids,HmdbEntry-method}
\alias{biofluids}
\title{extract biofluid associations}
\usage{
\S4method{biofluids}{HmdbEntry}(x)
}
\arguments{
\item{x}{HmdbEntry instance}
}
\value{
character vector
}
\description{
extract biofluid associations
}
\examples{
data(hmdb1)
biofluids(hmdb1)
}
