% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4.R
\name{biospecimens,HmdbEntry-method}
\alias{biospecimens,HmdbEntry-method}
\alias{biospecimens}
\title{extract biospecimen associations}
\usage{
\S4method{biospecimens}{HmdbEntry}(x)
}
\arguments{
\item{x}{HmdbEntry instance}
}
\value{
character vector
}
\description{
extract biospecimen associations
}
\examples{
data(hmdb1)
biospecimens(hmdb1)
}
