% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hmdb1}
\alias{hmdb1}
\title{hmdb1: demonstration HMDB entry}
\format{HmdbEntry instance}
\source{
www.hmdb.ca
}
\usage{
hmdb1
}
\description{
hmdb1: demonstration HMDB entry
}
\keyword{datasets}
