% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hmdb_disease}
\alias{hmdb_disease}
\title{hmdb_disease: Sept 2017 extract from hmdb.ca of direct disease associations}
\format{S4Vectors DataFrame}
\source{
www.hmdb.ca xml
}
\usage{
hmdb_disease
}
\description{
hmdb_disease: Sept 2017 extract from hmdb.ca of direct disease associations
}
\note{
This table also includes a column of metabolite name.
}
\keyword{datasets}
