% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hmdb_gene}
\alias{hmdb_gene}
\title{hmdb_gene: Sept 2017 extract from hmdb.ca of direct gene associations}
\format{S4Vectors DataFrame}
\source{
www.hmdb.ca xml
}
\usage{
hmdb_gene
}
\description{
hmdb_gene: Sept 2017 extract from hmdb.ca of direct gene associations
}
\keyword{datasets}
