% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hmdb_omim}
\alias{hmdb_omim}
\title{hmdb_omim: Sept 2017 extract from hmdb.ca of direct omim associations}
\format{
S4Vectors DataFrame
}
\source{
www.hmdb.ca xml
}
\usage{
hmdb_omim
}
\description{
hmdb_omim: Sept 2017 extract from hmdb.ca of direct omim associations
}
\keyword{datasets}
