% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hmdb_protein}
\alias{hmdb_protein}
\title{hmdb_protein: Sept 2017 extract from hmdb.ca of direct protein associations}
\format{
S4Vectors DataFrame
}
\source{
www.hmdb.ca xml
}
\usage{
hmdb_protein
}
\description{
hmdb_protein: Sept 2017 extract from hmdb.ca of direct protein associations
}
\keyword{datasets}
