% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4.R
\name{HmdbEntry}
\alias{HmdbEntry}
\title{Constructor for HmdbEntry instance}
\usage{
HmdbEntry(
  prefix = "http://www.hmdb.ca/metabolites/",
  id = "HMDB0000001",
  keepFull = TRUE
)
}
\arguments{
\item{prefix}{character(1) URL of HMDB source accepting queries for XML documents}

\item{id}{character(1) HMDB identifier tag}

\item{keepFull}{logical(1) indicating that entire parsed XML will be retained}
}
\value{
instance of HmdbEntry, or a list
}
\description{
Constructor for HmdbEntry instance
}
\note{
The XML returned by hmdb.ca can have different structures
for different metabolites.  If the mapping form XML to list
is not as anticipated for a given metabolite, the xmlToList
result is returned with a warning.  Such entries should be
reported to the hmdbQuery maintainer for map revision.
}
\examples{
HmdbEntry()
}
