% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4.R
\docType{methods}
\name{store,HmdbEntry-method}
\alias{store,HmdbEntry-method}
\alias{store}
\title{extract general association metadata in store slot}
\usage{
\S4method{store}{HmdbEntry}(x)
}
\arguments{
\item{x}{HmdbEntry instance}
}
\value{
list
}
\description{
extract general association metadata in store slot
}
\examples{
data(hmdb1)
names(store(hmdb1))
}
