% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4.R
\name{tissues,HmdbEntry-method}
\alias{tissues,HmdbEntry-method}
\alias{tissues}
\title{extract tissue associations}
\usage{
\S4method{tissues}{HmdbEntry}(x)
}
\arguments{
\item{x}{HmdbEntry instance}
}
\value{
character vector
}
\description{
extract tissue associations
}
\examples{
data(hmdb1)
tissues(hmdb1)
}
