% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4.R
\name{HmdbEntry}
\alias{HmdbEntry}
\title{Constructor for HmdbEntry instance}
\usage{
HmdbEntry(prefix = "http://www.hmdb.ca/metabolites/", id = "HMDB0000001",
  keepFull = TRUE)
}
\arguments{
\item{prefix}{character(1) URL of HMDB source accepting queries for XML documents}

\item{id}{character(1) HMDB identifier tag}

\item{keepFull}{logical(1) indicating that entire parsed XML will be retained}
}
\value{
instance of HmdbEntry
}
\description{
Constructor for HmdbEntry instance
}
\examples{
HmdbEntry()
}
